refactor: Streamline pipeline architecture for PRISM Zarr data
Major simplification of the climate data pipeline:

REMOVED:
- Eliminated preprocessor.py entirely (433 lines) - PRISM data is already clean
- Removed multi-format support and pattern matching
- Removed unnecessary data validation and unit conversions
- Removed coordinate standardization and outlier detection

SIMPLIFIED:
- config.py: 258 → 104 lines (60% reduction)
- data_loader.py: 338 → 154 lines (54% reduction)
- pipeline.py: 400+ → 328 lines (18% reduction)
- Total codebase: 1429 → 586 lines (59% reduction)

ARCHITECTURE:
Before: Config → Loader → Preprocessor → Indices → Output (5 phases)
After:  Config → Loader → Indices → Output (3 phases)

KEY IMPROVEMENTS:
- Direct PRISM Zarr loading without intermediate steps
- Automatic handling of external drive mount point changes
- Maintained all 84 climate indices functionality
- Cleaner, more maintainable codebase
- Faster execution with less overhead

This refactor leverages the fact that PRISM data is already:
- In correct units (Celsius for temp, mm for precipitation)
- Using standard dimensions (lat, lon, time)
- Clean and validated
- Properly structured in Zarr format
